+++ Changelog.md
@@ -13,9 +13,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Add `path` module as new args in `configlog` function at `app.js`. Completed on 2024-05-15
+
 ### Changed
 
 - Apply `html` instead `text` property in handler.webview.options. Completed on 2024-05-14
+- The `configlog` function in `app.js` will create a folder based on the engine name and save the success.log file in that folder. Completed on 2024-05-15
 
 ### Deprecated